making all template returns 'Manual'
Previously, some returns had explicit types and return statements which, due to the compiler, could not be avoided. These functions were therefore incompatible with Aborts and Failures, and would 'choke the pipeline'. It also meant that errors had to be communicated to the user (or a middle man) via codes with differing types!

Now that all returns are 'Manual', any symbol can be returned in place of the expected return, and so we can standardise error messages