all non-evaluated function calls throw error
Previously, a call to a function which did not match any pattern definition went silently unevaluated - this fact was often hidden by a trailing semicolon;
For example, below...
a = CreateQureg[3];
InitPlusState[A];
the second statement remains silently unevalauted, and qureg a is unchanged.

Now, any API function call which doesn't match a definition pattern will result in a generic erorr (instruct the user to read the doc of the function). This is significantly safer and alleviates a major user pitfall of QuESTlink.