patched delayed-eval Aliases
Previously, the RHS of a rule in Aliases was evaluated right BEFORE an attempted substitution was made. This caused some replacements (e.g. Table[ f[targ], {targ, {t}]) to prematurely evaluate (e.g. to f[{t}]), and hence substitute an unintended expression.

Ironically, this early evaluation occurred from an effort to delay evaluation when the user used a Rule instead of a RuleDelayed. Go figure!