patching Matr in GetCircuitInverse
Previously, Matr was assumed to be a unitary matrix like U and hence was conjugate-transposed, when it should instead be treated as an arbitrary matrix and hence directly inverted. Note that invertibility of the matrix is not explicitly asserted to simplify the result, much like unitarity of the U matrix being not yet asserted.