renamed RetargetCircuit to GetCircuitRetargeted
This is because the function is 'cheap' scaling linearly in-time with the number of gates in the input circuit, and so should belong to the 'Get' prefixed functions. Only functions which can prove unexpectedly expensive (e.g. SimplifyCircuit, RecompileCircuit, etc) are permitted their own prefixes. The new name is now more in-line with existing functions like 'GetCircuitCompacted' and 'GetCircuitGeneralised'. Frankly I'm unsure why I ever named it 'RetargetcCircuit', d'oh!