patched numeric Pauli string functions
which previously did not accept symbolically-expressed but ultimately numerical Pauli-string coefficients, like "Rational[1,10]", "Sqrt[2]" or "Pi". Instead, they catastrophically crashed the QuEST process! This is despite passing validation because they were recognised as numeric, although they were not instantiated as floating-point types necessary for passing to the C process.

Now, they are correctly explicitly cast to floats before being dispatched to the backend.

The affected functions are all which numerically dispatch Pauli strings to the backend:
- CalcExpecPauliStringDerivs
- SetQuregToPauliString
- CalcExpecPauliString
- ApplyPauliString
- CalcPauliStringMatrix
- SampleExpecPauliString

What a doozie!