patched DrawPauliTransferMap labelling
Previously, "PauliStringForm" values "String" and "Kronecker" did not heed the target qubits of the drawn map. This meant the drawn graph nodes were labelled as Pauli strings which targeted the least significant qubits. That is, the "internal, relative" states of the map were labelled, rather than of the targeted qubits